replace MessageNew, MessageEvent on the generic Event attribute use Event(Events::MESSAGE_NEW)
